2.1.0: cpu/image_mixer: Added support for image conversion and blending. Only interlacing support is missing.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.1.0@2314 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ accelerator/cpu/util/blend.h
@@ -35,7 +35,7 @@ static void blend(uint8_t* dest, const uint8_t* source1, const uint8_t* source2,
     const xmm_epi16 round  = 128;
     const xmm_epi16 lomask = 0x00FF;
 		
-	tbb::parallel_for(tbb::blocked_range<int>(0, count/sizeof(xmm_epi8)), [&](const tbb::blocked_range<int>& r)
+	tbb::parallel_for(tbb::blocked_range<size_t>(0, count/sizeof(xmm_epi8)), [&](const tbb::blocked_range<size_t>& r)
 	{
 		for(auto n = r.begin(); n != r.end(); ++n)    
 		{
